+++ src/kspdg/agent_api/example_agent.py
@@ -20,6 +20,7 @@ class NaivePursuitAgent(KSPDGBaseAgent):
 
         return {
             "burn_vec": [1.0, 0, 0, 1.0], # throttle in x-axis, throttle in y-axis, throttle in z-axis, duration [s]
+            "vec_type": 0,  # burn vector as throttle values (if =1, burn vector represents thrust in [N])
             "ref_frame": 0  # burn_vec expressed in agent vessel's right-handed body frame. 
                             # i.e. forward throttle, right throttle, down throttle, 
                             # Can also use rhcbci (1) and rhntw (2) ref frames
@@ -37,7 +38,7 @@ class PassivePursuitAgent(KSPDGBaseAgent):
         an abstract method
         """
 
-        return {"burn_vec": [0, 0, 0, 1.0], "ref_frame": 0}
+        return {"burn_vec": [0, 0, 0, 1.0], "vec_type": 0, "ref_frame": 0}
     
 class RandomPursuitAgent(KSPDGBaseAgent):
     """An agent that randomly burns in any direction"""
@@ -54,7 +55,7 @@ class RandomPursuitAgent(KSPDGBaseAgent):
         rnd_thr = np.random.uniform(-1, 1, 3)  # Generate random body-axes throttle values
         rnd_ctrl = np.append(rnd_thr, np.random.uniform(0, 2))  # append random duration
 
-        return {"burn_vec": list(rnd_ctrl), "ref_frame": 0}
+        return {"burn_vec": list(rnd_ctrl), "vec_type": 0, "ref_frame": 0}
     
 class ProgradePursuitAgent(KSPDGBaseAgent):
     """An agent that just burns prograde"""
@@ -68,6 +69,9 @@ class ProgradePursuitAgent(KSPDGBaseAgent):
         an abstract method
         """
 
-        return {"burn_vec": [0, 1.0, 0, 1.0], "ref_frame": 2}
+        return {
+            "burn_vec": [0, 8000.0, 0, 1.0], 
+            "vec_type": 1,  # burn vector represents thrust in [N]
+            "ref_frame": 2}